+++ spotinst_sdk2/test/elastigroup/aws/test_aws_eg_client.py
@@ -177,6 +177,20 @@ class AwsInitTestElastigroup(AwsInitTestCase):
 
 		self.assertEqual(response, self.mock_ok_res['response']['status'])
 
+	@patch('requests.get')
+	def testGetElastilog(self, mock):
+		mock_get_group_res    = self.load_json('../../test_lib/output/elastigroup/get_group_res.json')
+
+		self.mock_api_call.content = SimpleNamespace(**self.mock_api_call.content)
+		self.mock_api_call.content.decode = lambda code: json.dumps(mock_get_group_res)
+
+		mock.return_value = self.mock_api_call
+
+		response = self.client.get_elastilog(group_id="group_id", from_date="from_date", to_date="to_date", 
+			severity="severity", resource_id="resource_id", limit="limit")
+
+		self.assertEqual(len(response), len(mock_get_group_res["response"]["items"]))
+
 # Stateful Tests
 class AwsInitTestStateFul(AwsInitTestCase):