+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 All notable changes to this project will be documented in this file.
 This project adheres to [Semantic Versioning](http://semver.org/).
 
+## [2.1.2] - 2019-04-02
+### Added
+ - get_elastilog for aws client
+ - hpcGridEngine to azure eg model
+
 ## [2.1.1] - 2019-02-29
 ### Added
  - last 13 Azure endpoints